+++ docs/source/api/_server_interface.rst
@@ -1,0 +1,10 @@
+.. _server_interface:
+
+Server Interface
+================
+
+Common interface to the ESAF and Proposal information servers.
+
+.. automodule:: apsbss.server_interface
+    :members:
+
