--- docs/source/api/_servers.rst
@@ -1,10 +1,0 @@
-.. _servers:
-
-Servers
-=======
-
-Common interface to the  ESAF and Proposal information servers.
-
-.. automodule:: apsbss.servers
-    :members:
-
